4-13

+++ dubbo-bs/dubbo-provider/src/main/java/com/dubbo/provider/service/CommentServiceImpl.java
@@ -20,30 +20,44 @@ public class CommentServiceImpl implements CommentService{
     CommentMapper commentMapper;
     @Autowired
     private RedissonClient redissonClient;
-    public List<Comment> findAll(String articleid){return commentMapper.findAll(articleid);}
-    public void insertcomment(String articleid, String name, String content, String createtime, String state){ commentMapper.insertcomment(articleid,name,content,createtime,state);}
-    public void insertcomment1( String comment_id,String name,String replyname,String content,String prase_count,String createtime){commentMapper.insertcomment1(comment_id,name,replyname,content,prase_count,createtime);}
+    public List<Comment> findAll(String articleid){
+        if(articleid!=null)
+            return commentMapper.findAll(articleid);
+        else return null;
+    }
+    public void insertcomment(String articleid, String name, String content, String createtime, String state){
+        if(articleid!=null && name!=null && content!=null)
+             commentMapper.insertcomment(articleid,name,content,createtime,state);
+    }
+    public void insertcomment1( String comment_id,String name,String replyname,String content,String prase_count,String createtime){
+        if(comment_id!=null && name!=null && content!=null)
+              commentMapper.insertcomment1(comment_id,name,replyname,content,prase_count,createtime);
+    }
     public boolean updatesh(String id,String state){
-        RLock lock = redissonClient.getLock(id);
-        try {
-            boolean res = lock.tryLock(100, 10, TimeUnit.SECONDS);
-            if (!res) return false;
-            System.out.println("加锁成功！");
-            commentMapper.updatesh(id, state);
-        }catch (Exception e){
-            e.printStackTrace();
-            return false;
-        }finally {
-            lock.unlock();
-            System.out.println("解锁成功！");
-        }
-        return true;
+        if(id!=null) {
+            RLock lock = redissonClient.getLock(id);
+            try {
+                boolean res = lock.tryLock(100, 10, TimeUnit.SECONDS);
+                if (!res) return false;
+                System.out.println("加锁成功！");
+                commentMapper.updatesh(id, state);
+            } catch (Exception e) {
+                e.printStackTrace();
+                return false;
+            } finally {
+                lock.unlock();
+                System.out.println("解锁成功！");
+            }
+            return true;
+        }else return false;
     }
 
     @Transactional(rollbackFor = {RuntimeException.class, Error.class})
     public void deletecomment(String id){
-        commentMapper.deletecomment(id);
-        commentMapper.deletecomment1(id);
+        if(id!=null) {
+            commentMapper.deletecomment(id);
+            commentMapper.deletecomment1(id);
+        }
     }
 
 }
