+++ dubbo-bs/dubbo-provider/src/main/java/com/dubbo/provider/service/SearchContentServiceImpl.java
@@ -17,37 +17,58 @@ import java.util.List;
 public class SearchContentServiceImpl implements SearchContentService {
     @Resource
     SearchContentMapper searchContentMapper;
-    public List<SearchContent> findAll(String id){return searchContentMapper.findAll(id);}
+    public List<SearchContent> findAll(String id){
+        return searchContentMapper.findAll(id);
+    }
     public void insertzj(String iid,String zj,String title){
-         searchContentMapper.insertzj(iid,zj,title);
+        if(iid  != null && zj != null && title != null)
+            searchContentMapper.insertzj(iid,zj,title);
     }
     public void insertzj1(SearchContent1 searchContent1){
-        searchContentMapper.insertzj1(searchContent1);
+        if(searchContent1 != null)
+           searchContentMapper.insertzj1(searchContent1);
     }
-    public void updatezj(String id,String title){ searchContentMapper.updatezj(id,title); }
-    public void updatezj1(String id,String title,String url){ searchContentMapper.updatezj1(id,title,url); }
+    public void updatezj(String id,String title){
+        if(id != null && title != null)
+           searchContentMapper.updatezj(id,title);
+    }
+    public void updatezj1(String id,String title,String url){
+        if(id != null && title != null && url != null)
+           searchContentMapper.updatezj1(id,title,url);
+    }
 
     @Transactional(rollbackFor = {RuntimeException.class, Error.class})
     public void deletezj1(String id,String iid,int zj){
-        searchContentMapper.deletezj1(id);
-        for(SearchContent a:findAll(iid)){
-            if((a.zj==zj)&&(a.searchContent1.size()==0)){
-                deletesearchcontent(iid,String.valueOf(zj));
-                int i=0;
-                for(SearchContent c:findAll(iid)){
-                    updatesearchcontent(String.valueOf(c.id),String.valueOf(i+1));
-                    int j=0;
-                    for(SearchContent1 d:c.searchContent1){
-                        updatesearchcontent1(String.valueOf(d.id),String.valueOf(i+1),String.valueOf(i+1)+"-"+(j+1));
-                        j++;
+        if(id != null && iid != null && (Integer)zj !=null) {
+            searchContentMapper.deletezj1(id);
+            for (SearchContent a : findAll(iid)) {
+                if ((a.zj == zj) && (a.searchContent1.size() == 0)) {
+                    deletesearchcontent(iid, String.valueOf(zj));
+                    int i = 0;
+                    for (SearchContent c : findAll(iid)) {
+                        updatesearchcontent(String.valueOf(c.id), String.valueOf(i + 1));
+                        int j = 0;
+                        for (SearchContent1 d : c.searchContent1) {
+                            updatesearchcontent1(String.valueOf(d.id), String.valueOf(i + 1), String.valueOf(i + 1) + "-" + (j + 1));
+                            j++;
+                        }
+                        i++;
                     }
-                    i++;
                 }
             }
         }
     }
-    public void deletesearchcontent(String iid,String zj){searchContentMapper.deletesearchcontent(iid,zj);}
-    public void updatesearchcontent(String id,String zj){searchContentMapper.updatesearchcontent(id,zj);}
-    public void updatesearchcontent1(String id,String zj,String zjj){searchContentMapper.updatesearchcontent1(id,zj,zjj);}
+    public void deletesearchcontent(String iid,String zj){
+        if(iid != null && zj != null)
+           searchContentMapper.deletesearchcontent(iid,zj);
+    }
+    public void updatesearchcontent(String id,String zj){
+        if(id != null && zj != null)
+           searchContentMapper.updatesearchcontent(id,zj);
+    }
+    public void updatesearchcontent1(String id,String zj,String zjj){
+        if(id != null && zj != null && zjj != null)
+           searchContentMapper.updatesearchcontent1(id,zj,zjj);
+    }
 
 }
