+++ dubbo-bs/dubbo-provider/src/main/java/com/dubbo/provider/service/SearchServiceImpl.java
@@ -2,16 +2,14 @@ package com.dubbo.provider.service;
 
 
 import com.dubbo.api.model.Search;
-import com.dubbo.api.service.CommentService;
 import com.dubbo.api.service.SearchService;
 import com.dubbo.provider.mapper.SearchMapper;
-import javafx.beans.binding.BooleanExpression;
+import com.dubbo.provider.util.RedisLockCommon;
 import org.redisson.api.RLock;
 import org.redisson.api.RedissonClient;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.ResponseBody;
 
 import javax.annotation.Resource;
 import java.util.List;
@@ -32,39 +30,43 @@ public class SearchServiceImpl implements SearchService {
     }
     public List<Search> findById(String id,String author){return searchMapper.findById(id,author);}
     public Boolean updateimg(Search search){
-        String searchId=""+search.getId();
-
-
-        RLock lock = redissonClient.getLock(searchId);
-
-
-        try{
-            boolean res=lock.tryLock(100,10,TimeUnit.SECONDS);
-            if(!res) return false;
-            System.out.println("加锁成功！");
-        searchMapper.updateimg(search);
-        }catch (Exception e){
-            e.printStackTrace();
-            return false;
-        }finally {
-            lock.unlock();
-            redisTemplate.opsForZSet().removeRange("page",0,-1);
-            search2(null,null);
-            System.out.println("解锁成功！");
-        }
-        return true;
+        if(search != null) {
+            String searchId = "" + search.getId();
+            RLock lock = redissonClient.getLock(searchId);
+            try {
+                boolean res = lock.tryLock(100, 10, TimeUnit.SECONDS);
+                if (!res) return false;
+                System.out.println("加锁成功！");
+                searchMapper.updateimg(search);
+            } catch (Exception e) {
+                e.printStackTrace();
+                return false;
+            } finally {
+                lock.unlock();
+                redisTemplate.opsForZSet().removeRange("page", 0, -1);
+                search2(null, null);
+                System.out.println("解锁成功！");
+            }
+            return true;
+        }else return false;
     }
     public void insertlesson(Search search){
-        searchMapper.insertlesson(search);
-        redisTemplate.opsForZSet().removeRange("page",0,-1);
-        search2(null,null);
+        if(search != null) {
+            searchMapper.insertlesson(search);
+            redisTemplate.opsForZSet().removeRange("page", 0, -1);
+            search2(null, null);
+        }
     }
     public void deletelesson(String id){
-        searchMapper.deletelesson(id);
-        redisTemplate.opsForZSet().removeRange("page",0,-1);
-        search2(null,null);
+        if(id != null) {
+            searchMapper.deletelesson(id);
+            redisTemplate.opsForZSet().removeRange("page", 0, -1);
+            search2(null, null);
+        }
     }
-    public List<Search> getHotsearch(){ return  searchMapper.getHotsearch();}
+    public List<Search> getHotsearch(){
+        return searchMapper.getHotsearch();
+    }
     public Object search1(String a,String b){
         if(a==null||b==null)
             return redisTemplate.opsForZSet().range("page",0,-1);
@@ -83,9 +85,15 @@ public class SearchServiceImpl implements SearchService {
             return redisTemplate.opsForZSet().range("page", Long.parseLong(a), Long.parseLong(b) + Long.parseLong(a) - 1);
         }
     }
-    public List<Search> search2(int url){return searchMapper.search2(url);}
+    public List<Search> search2(int url){
+        if((Integer)url != null) return searchMapper.search2(url);
+        else return null;
+    }
     public int selectBestId(){return searchMapper.selectBestId();}
     public int selectBestId1(){return searchMapper.selectBestId1();}
-    public Object findAll1(String start,String num){return searchMapper.findAll1(start,num);}
+    public Object findAll1(String start,String num){
+        if(start !=null && num != null) return searchMapper.findAll1(start,num);
+        else return null;
+    }
 
 }
